Made society slice

+++ src/pages/society/Societyroot.tsx
@@ -3,8 +3,21 @@ import { toast } from "sonner";
 import Cookies from "universal-cookie";
 import { Button } from "@/components/ui/button";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import useSocietyProfile from "@/hooks/useSocietyProfile";
 
 const Societyroot: React.FC = () => {
+  useSocietyProfile();
   const cookies = new Cookies(null, { path: "/" });
   const navigate = useNavigate();
   return (
@@ -52,20 +65,38 @@ const Societyroot: React.FC = () => {
           Edit Profile
         </NavLink>
 
-        <Button
-          onClick={() => {
-            cookies.remove("token");
-            localStorage.clear();
-            // dispatch(deleteUser());
-            navigate("/");
-            toast("Logout successful", {
-              description: "You have successfully logged out.",
-            });
-          }}
-          className="text-sm  min-w-[80%] max-w-[300px] border border-red-500 text-red-500 bg-transparent  hover:bg-red-500 hover:text-white transition-all duration-300"
-        >
-          Logout
-        </Button>
+        <AlertDialog>
+          <AlertDialogTrigger className="min-w-[80%] max-w-[300px]">
+            <Button className="text-sm  min-w-[80%] max-w-[300px] border border-red-500 text-red-500 bg-transparent  hover:bg-red-500 hover:text-white transition-all duration-300">
+              Logout
+            </Button>
+          </AlertDialogTrigger>
+          <AlertDialogContent>
+            <AlertDialogHeader>
+              <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+              <AlertDialogDescription>
+                This action cannot be undone. You will be logged out of
+                ConnectHub.
+              </AlertDialogDescription>
+            </AlertDialogHeader>
+            <AlertDialogFooter>
+              <AlertDialogCancel>Cancel</AlertDialogCancel>
+              <AlertDialogAction
+                onClick={() => {
+                  cookies.remove("token");
+                  localStorage.clear();
+                  // dispatch(deleteUser());
+                  navigate("/");
+                  toast("Logout successful", {
+                    description: "You have successfully logged out.",
+                  });
+                }}
+              >
+                Continue
+              </AlertDialogAction>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialog>
       </div>
       <Outlet />
     </div>
